backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
groq
python-dotenv
